Latest version 2.6 of the FQLite:
Added support for databases with a size > 2,1 GB
Hex-Viewer is now only displayed if needed (not allways in Background)
There is a new ViewLog option in the 'Info'-Menu - you can now check the
log-messages
Columns for header fields, schema-info  and pages-tab are new sortable
Filter function on table view has been updated
Links in the Schema-Detail-View now work as expected
Fonts-Dialog now offers Support to go back to Default Font
Smaller bug fixes
Improved support for decoding BLOB-columns (BSON, MessagePack, AVRO,
Thrift...)
Improved detection of deleted files
New pages tab for preview page types
Improved Filter-function for tables (you can now filter for a particular
column)
Support for pdf-BLOBs with internal PDF-Viewer
copy Tooltip content to clipboard
view .tiff and .heic BLOBs with OS default viewer fort those file types
clicking on the <help> - function will open an PDF-Version from the user
manual (before, the online help was displayed)

+++ src/fqlite/util/CarvingResult.java
@@ -13,10 +13,10 @@ public class CarvingResult {
 
 	public StringBuffer bf;
 	public int rcursor;
-	public int offset; 
+	public long offset; 
 	public LinkedList<String> record;
 	
-	public CarvingResult(int rcursor,int offset, StringBuffer result, LinkedList<String> record)
+	public CarvingResult(int rcursor,long offset, StringBuffer result, LinkedList<String> record)
 	{
 		bf = result;
 		this.rcursor = rcursor;
